Changes the sensor code to hit the 100Hz mark.
In the nominal mode, measureCriticalSensors is called every iteration of the AQ 'loop' function. This function then measures gyro and acc which takes about 0.6 ms (1440Hz). If this happens as frequently as possible then the 100Hz cycle happens at 96.7Hz because of the delay incurred within 0.6 ms of the 100Hz step (e.g. statistically a 0.6ms measurement occurs 0.3 ms before the 100Hz counter, which results in a delay of 0.3 ms each step. reducing the frequency by 3Hz).

This changeset alters the function measureCriticalSensors to prevent a measurement within 0.5 ms of the start of the 100Hz step *only from the AQ_v2 board because that's what we have and that's where I measure 0.6ms for the measurement cycle. The result of this protection is that the 100Hz cycle occurs at 100.00+-0.02 Hz, 50Hz occurs at 50.00+-0.01 Hz, and the 10HZ at 10.00 +- 0.01Hz. The downside of this change is that you lose 1-2 measurements (out of 20) on the gyro and mag.

That the measurement you lose is the one closest to the 100Hz cycle start shouldn't matter as it's not weighted any differently to the other 20 measurements.